[#165] chore: 커스텀 styled components 수정
