+++ client/src/lib/customStyledComponents/index.tsx
@@ -1,43 +1,15 @@
-import { useEffect } from 'react';
+/* eslint-disable @typescript-eslint/ban-types */
+import constructWithTag from '~/lib/customStyledComponents/constructWithTag';
+import domElements from './utils/domElements';
 
-import applyToHead from './utils/applyToHead';
-import checkIsValidDOMProps from './utils/checkIsValidDOMProps';
-import cssParser from './utils/cssParser';
-import cssSerializer from './utils/cssStringify';
-import { generateAlphabeticName } from './utils/generateAlphaName';
+interface Styled {
+  [tag: string]: (string: TemplateStringsArray, ...args: unknown[]) => Function;
+}
 
-let classNameIndex = 1000;
+const styled: Styled = {};
 
-export type PropsType = { [key: string]: any };
+domElements.forEach((domElement) => {
+  styled[domElement] = constructWithTag(domElement);
+});
 
-/**
- * Tag를 받아 Component 클로져를 반환합니다.
- */
-const constructWithTag = (tag: string) => {
-  const Tag = `${tag}` as keyof JSX.IntrinsicElements;
-
-  const construct =
-    (templateStr: TemplateStringsArray, ...args: unknown[]) =>
-    (props: PropsType) => {
-      const className = generateAlphabeticName(++classNameIndex);
-      const classNameWithSuffix = `custom-sc-${className}`;
-
-      useEffect(() => {
-        const parsedCSS = cssParser(templateStr, props, args);
-        const serializedString = cssSerializer(tag, parsedCSS);
-        applyToHead(serializedString);
-      }, [props]);
-
-      const domProps = checkIsValidDOMProps(tag, props);
-
-      return (
-        <Tag {...domProps} className={classNameWithSuffix}>
-          {props.children}
-        </Tag>
-      );
-    };
-
-  return construct;
-};
-
-export default constructWithTag;
+export default styled;
